requirements.txt
# Rasa
rasa==3.6.21
rasa-sdk==3.6.2

# Audio
pvporcupine==3.0.5
pvrecorder==1.2.7
sounddevice==0.5.2
scipy==1.10.1
numpy==1.23.5
openai-whisper==20250625
torch>=2.1.0
requests==2.32.5
